[CSM Portal Microapp] Relocate MetaRow from customers/ to common/
Now used by Security Center's vulnerability detail page too, not just
Customers — components/customers/MetaRow implied a scope it no longer has.
